seccure delete video & real time processing using wire:poll

+++ app/Http/Livewire/Video/AllVideo.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Livewire\Video;
 
+use App\Models\Channel;
 use App\Models\Video;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\Storage;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -10,23 +12,33 @@ use Livewire\WithPagination;
 class AllVideo extends Component
 {
     use WithPagination;
+    use AuthorizesRequests;
     protected $paginationTheme = 'bootstrap';
+
+    public $channel;
+
+    public function mount(Channel $channel)
+    {
+        $this->channel = $channel;
+    }
 
     public function render()
     {
         return view('livewire.video.all-video')
-        ->with('videos', auth()->user()->channel->videos()->paginate(1))
+        ->with('videos', $this->channel->videos()->paginate(3))
         ->extends('layouts.app');
 
     }
 
     public function delete(Video $video)
     {
+        // if user is allowed to delete a video
+        $this->authorize('delete',$video);
         $deleted = Storage::disk('videos')->deleteDirectory($video->uid);
         if($deleted){
             $video->delete();
         }
-        return back(); 
+        return back();
     }
 
 }
